requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
bcrypt==5.0.0
billiard==4.2.4
bleach==6.3.0
cbor2==5.8.0
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
dnspython==2.8.0
drf-yasg==1.21.11
ecdsa==0.19.1
email-validator==2.3.0
eventlet==0.40.4
exceptiongroup==1.3.1
fastapi==0.121.2
frozenlist==1.8.0
gevent==25.9.1
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
kombu==5.6.1
lxml==6.0.2
msgpack==1.1.2
multidict==6.7.0
packaging==25.0
passlib==1.7.4
phonenumbers==9.0.22
pillow==12.0.0
prompt_toolkit==3.0.52
propcache==0.4.1
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
PyMySQL==1.1.2
pyOpenSSL==25.3.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
redis==7.1.0
reportlab==4.4.6
requests==2.32.5
rsa==4.9.1
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.49.3
twilio==9.9.1
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
user-agents==2.2.0
uvicorn==0.38.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
webencodings==0.5.1
websockets==15.0.1
yarl==1.22.0
zope.event==6.1
zope.interface==8.1.1
